backend updates, cleanup
