+++ fiftyone/core/state.py
@@ -9,7 +9,6 @@ Defines the shared state between the FiftyOne App and backend.
 import logging
 import typing as t
 
-from bson import json_util
 from dataclasses import asdict
 from mongoengine.base import BaseDict, BaseList
 import strawberry as gql
@@ -23,7 +22,6 @@ from fiftyone.core.config import AppConfig
 import fiftyone.core.dataset as fod
 from fiftyone.core.odm.dataset import ColorScheme
 from fiftyone.core.odm.workspace import Space
-import fiftyone.core.media as fom
 import fiftyone.core.odm as foo
 import fiftyone.core.utils as fou
 import fiftyone.core.view as fov
@@ -122,7 +120,7 @@ class StateDescription(etas.Serializable):
                 d["colorscale"] = self.config.get_colormap()
 
             if isinstance(self.spaces, Space):
-                d["spaces"] = self.spaces.to_json()
+                d["spaces"] = self.spaces.to_dict()
 
             if isinstance(self.color_scheme, ColorScheme):
                 d["color_scheme"] = self.color_scheme.to_dict(False)
@@ -182,7 +180,7 @@ class StateDescription(etas.Serializable):
 
         spaces = d.get("spaces", None)
         if spaces is not None:
-            spaces = Space.from_dict(json_util.loads(spaces))
+            spaces = Space.from_dict(spaces)
 
         color_scheme = d.get("color_scheme", None)
         if color_scheme:
@@ -246,9 +244,11 @@ def serialize_fields(schema: t.Dict) -> t.List[SampleField]:
                     embedded_doc_type=embedded_doc_type,
                     subfield=subfield,
                     description=field.description,
-                    info=_convert_mongoengine_data(field.info)
-                    if isinstance(field.info, BaseDict)
-                    else field.info,
+                    info=(
+                        _convert_mongoengine_data(field.info)
+                        if isinstance(field.info, BaseDict)
+                        else field.info
+                    ),
                 )
             )
