+++ fiftyone/server/routes/events.py
@@ -5,14 +5,16 @@ FiftyOne Server /events route
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
+
 import typing as t
 
 from starlette.endpoints import HTTPEndpoint
 from starlette.requests import Request
 from sse_starlette.sse import EventSourceResponse
 
+from fiftyone.core.session.events import ListenPayload
+
 from fiftyone.server.events import (
-    ListenPayload,
     add_event_listener,
     dispatch_polling_event_listener,
 )
