+++ fiftyone/server/routes/screenshot.py
@@ -5,6 +5,7 @@ FiftyOne Server /screenshot route
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
+
 import typing as t
 
 from starlette.endpoints import HTTPEndpoint
@@ -13,9 +14,9 @@ from starlette.responses import Response
 
 from fiftyone.core.session.events import get_screenshot, DeactivateNotebookCell
 from fiftyone.core.session.templates import SCREENSHOT_DATABRICKS
-from fiftyone.server.events import dispatch_event
 
 from fiftyone.server.decorators import route
+from fiftyone.server.events import dispatch_event
 
 
 class Screenshot(HTTPEndpoint):
