+++ fiftyone/server/routes/select.py
@@ -5,6 +5,7 @@ FiftyOne Server /select route
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
+
 from starlette.endpoints import HTTPEndpoint
 from starlette.requests import Request
 from fiftyone.core.session.events import StateUpdate
