+++ fiftyone/server/scalars.py
@@ -5,7 +5,8 @@ FiftyOne Server GraphQL scalars
 | `voxel51.com <https://voxel51.com/>`_
 |
 """
-from datetime import date, datetime, timezone
+
+from datetime import date, datetime
 from bson import json_util
 import json
 import strawberry as gql
